Refactor for command line utility
Split solution into three projects:
- One WinForm
- One Common
- One Command Line

The latter is empty as this sweep was meant to lay the foundation for the build of a command line utility
