+++ CommandLine/Program.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ImageToLego
+{
+  class CommandLineDriver
+  {
+    static void Main(string[] args)
+    {
+    }
+  }
+}
