+++ WinFormDriver/Properties/AssemblyInfo.cs
