+++ WinFormDriver/Properties/Settings.settings
