+++ WinFormDriver/WinFormDriver.csproj
@@ -53,13 +53,8 @@
     <Compile Include="Classes\CustomPanel.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Classes\LegoPiece.cs" />
-    <Compile Include="Classes\PaletteProcessor.cs" />
-    <Compile Include="Extensions\ColorExtensions.cs" />
-    <Compile Include="Helper\CSVTools.cs" />
     <Compile Include="Helper\ImageHelper.cs" />
     <Compile Include="Helper\TwoDimensionalArrayHelper.cs" />
-    <Compile Include="Helper\TypeCoversions.cs" />
     <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -97,6 +92,15 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="LEGO.ico" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Extensions\" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Shared\Shared.csproj">
+      <Project>{25d9b41a-563a-4d1f-8206-7bfd2f40f462}</Project>
+      <Name>Shared</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
