+++ colours.csv
@@ -1,5 +1,54 @@
-Name, R,G,B
-Red,255,0,0
-Green,0,255,0
-Blue,0,0,255
-Purple,127,0,255
+Name,R,G,B
+Aqua,181,211,214
+Black,0,0,0
+Blue,0,87,166
+Bright Light Orange,247,186,48
+Bright Light Yellow,243,224,85
+Bright Pink,255,187,255
+Brown,83,33,21
+Dark Azure,51,153,255
+Dark Blue,20,48,68
+Dark Brown,51,0,0
+Dark Gray,107,90,90
+Dark Green,46,85,67
+Dark Orange,179,84,8
+Dark Pink,200,112,128
+Dark Purple,95,38,131
+Dark Red,106,14,21
+Dark Stone Gray,89,93,96
+Dark Tan,144,116,80
+Dark Turquoise,0,138,128
+Flesh,204,128,98
+Green,0,100,46
+Lavender,203,162,200
+Light Aqua,204,255,255
+Light Blue,180,210,227
+Light Flesh,254,204,176
+Light Gray,156,156,156
+Light Orange,247,173,99
+Light Stone Gray,228,232,232
+Light Yellow,255,227,131
+Lime,166,202,85
+Maersk Blue,107,173,214
+Magenta,181,41,82
+Medium Azure,66,192,251
+Medium Blue,97,175,255
+Medium Dark Flesh,231,139,62
+Medium Dark Pink,247,133,177
+Medium Green,98,245,142
+Medium Lavender,181,165,213
+Medium Orange,255,165,49
+Medium Stone Gray,175,181,199
+Olive Green,124,144,81
+Orange,255,126,20
+Pearl Light Gray,172,183,192
+Pink,255,192,203
+Purple,165,73,156
+Red,179,0,6
+Reddish Brown,137,53,29
+Sand Blue,90,113,132
+Sand Green,118,162,144
+Tan,222,198,156
+White,255,255,255
+Yellow,247,209,23
+Yellowish Green,223,238,165
